+++ apostello/forms.py
@@ -5,12 +5,18 @@ from apostello.models import Keyword, Recipient, RecipientGroup, UserProfile
 from apostello.validators import gsm_validator, less_than_sms_char_limit
 
 
+def get_content_rows():
+    from math import ceil
+    from site_config.models import SiteConfiguration
+    return ceil(SiteConfiguration.get_solo().sms_char_limit / 160)
+
+
 class SendAdhocRecipientsForm(forms.Form):
     """Send an sms to ad-hoc groups."""
     content = forms.CharField(
         validators=[gsm_validator, less_than_sms_char_limit],
         required=True,
-        min_length=1
+        min_length=1,
     )
     recipients = forms.ModelMultipleChoiceField(
         queryset=Recipient.objects.filter(is_archived=False),
@@ -48,6 +54,7 @@ class SendAdhocRecipientsForm(forms.Form):
     def __init__(self, *args, **kwargs):
         self.user = kwargs.pop('user', None)
         super(SendAdhocRecipientsForm, self).__init__(*args, **kwargs)
+        self.fields['content'].widget = forms.Textarea(attrs={'rows': get_content_rows()})
 
 
 class SendRecipientGroupForm(forms.Form):
@@ -55,7 +62,7 @@ class SendRecipientGroupForm(forms.Form):
     content = forms.CharField(
         validators=[gsm_validator, less_than_sms_char_limit],
         required=True,
-        min_length=1
+        min_length=1,
     )
     recipient_group = forms.ModelChoiceField(
         queryset=RecipientGroup.objects.filter(is_archived=False),
@@ -92,6 +99,7 @@ class SendRecipientGroupForm(forms.Form):
     def __init__(self, *args, **kwargs):
         self.user = kwargs.pop('user', None)
         super(SendRecipientGroupForm, self).__init__(*args, **kwargs)
+        self.fields['content'].widget = forms.Textarea(attrs={'rows': get_content_rows()})
 
 
 class ManageRecipientGroupForm(forms.ModelForm):
@@ -258,3 +266,4 @@ class GroupAllCreateForm(forms.Form):
         'If this group already exists it will be overwritten.',
         max_length=150,
     )
+